+++ providers/tests/fab/auth_manager/api_endpoints/test_variable_endpoint.py
@@ -27,7 +27,6 @@ from tests_common.test_utils.db import clear_db_variables
 
 pytestmark = [
     pytest.mark.db_test,
-    pytest.mark.skip_if_database_isolation_mode,
     pytest.mark.skipif(not AIRFLOW_V_3_0_PLUS, reason="Test requires Airflow 3.0+"),
 ]