+++ providers/tests/google/marketing_platform/operators/test_campaign_manager.py
@@ -85,7 +85,6 @@ class TestGoogleCampaignManagerDeleteReportOperator:
         )
 
 
-@pytest.mark.skip_if_database_isolation_mode
 @pytest.mark.db_test
 class TestGoogleCampaignManagerDownloadReportOperator:
     def setup_method(self):