+++ tests/jobs/test_scheduler_job.py
@@ -98,7 +98,7 @@ from tests_common.test_utils.mock_operators import CustomOperator
 if AIRFLOW_V_3_0_PLUS:
     from airflow.utils.types import DagRunTriggeredByType
 
-pytestmark = [pytest.mark.db_test, pytest.mark.skip_if_database_isolation_mode]
+pytestmark = pytest.mark.db_test
 
 ROOT_FOLDER = os.path.realpath(
     os.path.join(os.path.dirname(os.path.realpath(__file__)), os.pardir, os.pardir)
@@ -181,6 +181,12 @@ class TestSchedulerJob:
         default_executor.name = ExecutorName(alias="default_exec", module_path="default.exec.module.path")
         second_executor = mock.MagicMock(name="SeconadaryExecutor", slots_available=8, slots_occupied=0)
         second_executor.name = ExecutorName(alias="secondary_exec", module_path="secondary.exec.module.path")
+
+        # TODO: Task-SDK Make it look like a bound method. Needed until we remove the old queue_command
+        # interface from executors
+        default_executor.queue_workload.__func__ = BaseExecutor.queue_workload
+        second_executor.queue_workload.__func__ = BaseExecutor.queue_workload
+
         with mock.patch("airflow.jobs.job.Job.executors", new_callable=PropertyMock) as executors_mock:
             executors_mock.return_value = [default_executor, second_executor]
             yield [default_executor, second_executor]
@@ -540,7 +546,7 @@ class TestSchedulerJob:
         self.job_runner._critical_section_enqueue_task_instances(session)
         session.flush()
         ti1.refresh_from_db(session=session)
-        assert State.SCHEDULED == ti1.state
+        assert ti1.state == State.SCHEDULED
         session.rollback()
 
     # @pytest.mark.usefixtures("mock_executor")
@@ -714,7 +720,7 @@ class TestSchedulerJob:
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
         session.flush()
-        assert 3 == len(res)
+        assert len(res) == 3
         res_keys = []
         for ti in res:
             res_keys.append(ti.key)
@@ -855,7 +861,7 @@ class TestSchedulerJob:
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-        assert 5 == len(res)
+        assert len(res) == 5
         res_ti_keys = [res_ti.key for res_ti in res]
         for ti in tis_tuple:
             assert ti.key in res_ti_keys
@@ -894,7 +900,7 @@ class TestSchedulerJob:
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-        assert 2 == len(res)
+        assert len(res) == 2
         assert ti3.key == res[0].key
         assert ti2.key == res[1].key
 
@@ -952,14 +958,14 @@ class TestSchedulerJob:
 
         # Two tasks w/o pool up for execution and our default pool size is 1
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
 
         ti2.state = State.RUNNING
         session.flush()
 
         # One task w/o pool up for execution and one task running
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
-        assert 0 == len(res)
+        assert len(res) == 0
 
         session.rollback()
         session.close()
@@ -989,7 +995,7 @@ class TestSchedulerJob:
         session.flush()
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
         session.flush()
-        assert 0 == len(res)
+        assert len(res) == 0
         tis = dr.get_task_instances(session=session)
         assert len(tis) == 2
         assert all(ti.state == State.FAILED for ti in tis)
@@ -1012,7 +1018,7 @@ class TestSchedulerJob:
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
         session.flush()
-        assert 0 == len(res)
+        assert len(res) == 0
         session.rollback()
 
     def test_infinite_pool(self, dag_maker):
@@ -1039,7 +1045,7 @@ class TestSchedulerJob:
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
         session.flush()
-        assert 1 == len(res)
+        assert len(res) == 1
         session.rollback()
 
     def test_not_enough_pool_slots(self, caplog, dag_maker):
@@ -1096,7 +1102,7 @@ class TestSchedulerJob:
         self.job_runner = SchedulerJobRunner(job=scheduler_job, subdir=os.devnull)
         session = settings.Session()
 
-        assert 0 == len(self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session))
+        assert len(self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)) == 0
         session.rollback()
 
     def test_tis_for_queued_dagruns_are_not_run(self, dag_maker):
@@ -1122,7 +1128,7 @@ class TestSchedulerJob:
         session.flush()
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-        assert 1 == len(res)
+        assert len(res) == 1
         assert ti2.key == res[0].key
         ti1.refresh_from_db()
         ti2.refresh_from_db()
@@ -1225,7 +1231,7 @@ class TestSchedulerJob:
             loader_mock.side_effect = executor.get_mock_loader_side_effect()
             res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-            assert 2 == len(res)
+            assert len(res) == 2
 
             ti1_1.state = State.RUNNING
             ti2.state = State.RUNNING
@@ -1238,7 +1244,7 @@ class TestSchedulerJob:
 
             res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-            assert 1 == len(res)
+            assert len(res) == 1
 
             ti1_2.state = State.RUNNING
             ti1_3 = dr3.get_task_instance(task1.task_id)
@@ -1249,7 +1255,7 @@ class TestSchedulerJob:
 
             res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-            assert 0 == len(res)
+            assert len(res) == 0
 
             ti1_1.state = State.SCHEDULED
             ti1_2.state = State.SCHEDULED
@@ -1261,7 +1267,7 @@ class TestSchedulerJob:
 
             res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-            assert 2 == len(res)
+            assert len(res) == 2
 
             ti1_1.state = State.RUNNING
             ti1_2.state = State.SCHEDULED
@@ -1273,7 +1279,7 @@ class TestSchedulerJob:
 
             res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-            assert 1 == len(res)
+            assert len(res) == 1
             session.rollback()
 
     def test_change_state_for_executable_task_instances_no_tis_with_state(self, dag_maker):
@@ -1303,7 +1309,7 @@ class TestSchedulerJob:
         session.flush()
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=100, session=session)
-        assert 0 == len(res)
+        assert len(res) == 0
 
         session.rollback()
 
@@ -1330,7 +1336,7 @@ class TestSchedulerJob:
         # Schedule ti with lower priority,
         # because the one with higher priority is limited by a concurrency limit
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
         assert res[0].key == ti2.key
 
         session.rollback()
@@ -1367,7 +1373,7 @@ class TestSchedulerJob:
         # Schedule ti with lower priority,
         # because the one with higher priority is limited by a concurrency limit
         res = self.job_runner._executable_task_instances_to_queued(max_tis=1, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
         assert res[0].key == ti2.key
 
         session.rollback()
@@ -1396,7 +1402,7 @@ class TestSchedulerJob:
         # Schedule ti with lower priority,
         # because the one with higher priority is limited by a concurrency limit
         res = self.job_runner._executable_task_instances_to_queued(max_tis=1, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
         assert res[0].key == ti1b.key
 
         session.rollback()
@@ -1425,7 +1431,7 @@ class TestSchedulerJob:
         # Schedule ti with higher priority,
         # because it's running in a different DAG run with 0 active tis
         res = self.job_runner._executable_task_instances_to_queued(max_tis=1, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
         assert res[0].key == ti2a.key
 
         session.rollback()
@@ -1458,7 +1464,7 @@ class TestSchedulerJob:
         # Schedule ti with lower priority,
         # because the one with higher priority is limited by a concurrency limit
         res = self.job_runner._executable_task_instances_to_queued(max_tis=1, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
         assert res[0].key == ti1b.key
 
         session.rollback()
@@ -1495,7 +1501,7 @@ class TestSchedulerJob:
         session.flush()
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=1, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
         assert res[0].key == ti1.key
 
         session.rollback()
@@ -1519,7 +1525,7 @@ class TestSchedulerJob:
         session.flush()
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
-        assert 0 == len(res)
+        assert len(res) == 0
 
         mock_stats_gauge.assert_has_calls(
             [
@@ -1535,7 +1541,7 @@ class TestSchedulerJob:
         session.flush()
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
-        assert 1 == len(res)
+        assert len(res) == 1
 
         mock_stats_gauge.assert_has_calls(
             [
@@ -1727,7 +1733,7 @@ class TestSchedulerJob:
         scheduler_job.max_tis_per_query = 6
         # First pass we'll grab 6 of the 8 tasks (limited by max_tis_per_query)
         res = self.job_runner._critical_section_enqueue_task_instances(session)
-        assert 6 == res
+        assert res == 6
         for ti in tis1[:3] + tis2[:3]:
             ti.refresh_from_db()
             assert ti.state == TaskInstanceState.QUEUED
@@ -1737,11 +1743,11 @@ class TestSchedulerJob:
 
         # The remaining TIs are queued
         res = self.job_runner._critical_section_enqueue_task_instances(session)
-        assert 2 == res
+        assert res == 2
 
         for ti in tis1 + tis2:
             ti.refresh_from_db()
-            assert State.QUEUED == ti.state
+            assert ti.state == State.QUEUED
 
     @pytest.mark.parametrize(
         "task1_exec, task2_exec",
@@ -1790,7 +1796,7 @@ class TestSchedulerJob:
         scheduler_job.max_tis_per_query = 2
 
         total_enqueued = self.job_runner._critical_section_enqueue_task_instances(session)
-        assert 2 == total_enqueued
+        assert total_enqueued == 2
 
     def test_execute_task_instances_limit_slots(self, dag_maker, mock_executors):
         dag_id = "SchedulerJobTest.test_execute_task_instances_limit"
@@ -2866,7 +2872,7 @@ class TestSchedulerJob:
 
         res = self.job_runner._executable_task_instances_to_queued(max_tis=32, session=session)
 
-        assert [] == res
+        assert res == []
 
     @pytest.mark.parametrize(
         "ti_states, run_state",
@@ -2950,7 +2956,7 @@ class TestSchedulerJob:
                 # zero tasks ran
                 assert len(session.query(TaskInstance).filter(TaskInstance.dag_id == dag_id).all()) == 0
                 session.commit()
-                assert [] == self.null_exec.sorted_tasks
+                assert self.null_exec.sorted_tasks == []
 
                 # previously, running this backfill would kick off the Scheduler
                 # because it would take the most recent run and start from there
@@ -2977,7 +2983,7 @@ class TestSchedulerJob:
                 # still one task
                 assert len(session.query(TaskInstance).filter(TaskInstance.dag_id == dag_id).all()) == 1
                 session.commit()
-                assert [] == self.null_exec.sorted_tasks
+                assert self.null_exec.sorted_tasks == []
 
     @pytest.mark.parametrize(
         "configs",
@@ -3581,7 +3587,7 @@ class TestSchedulerJob:
         scheduler_job = Job()
         self.job_runner = SchedulerJobRunner(job=scheduler_job)
         session = settings.Session()
-        assert 0 == self.job_runner.adopt_or_reset_orphaned_tasks(session=session)
+        assert self.job_runner.adopt_or_reset_orphaned_tasks(session=session) == 0
 
     @pytest.mark.parametrize(
         "adoptable_state",
@@ -3607,7 +3613,7 @@ class TestSchedulerJob:
         session.commit()
 
         num_reset_tis = self.job_runner.adopt_or_reset_orphaned_tasks(session=session)
-        assert 1 == num_reset_tis
+        assert num_reset_tis == 1
 
     def test_adopt_or_reset_orphaned_tasks_external_triggered_dag(self, dag_maker, session):
         dag_id = "test_reset_orphaned_tasks_external_triggered_dag"
@@ -3654,7 +3660,7 @@ class TestSchedulerJob:
         session.flush()
 
         assert dr1.run_type == DagRunType.BACKFILL_JOB
-        assert 0 == self.job_runner.adopt_or_reset_orphaned_tasks(session=session)
+        assert self.job_runner.adopt_or_reset_orphaned_tasks(session=session) == 0
         session.rollback()
 
     def test_reset_orphaned_tasks_no_orphans(self, dag_maker):
@@ -3698,13 +3704,13 @@ class TestSchedulerJob:
         dr1 = dag_maker.create_dagrun()
         dr1.state = State.QUEUED
         tis = dr1.get_task_instances(session=session)
-        assert 1 == len(tis)
+        assert len(tis) == 1
         tis[0].state = State.SCHEDULED
         session.merge(dr1)
         session.merge(tis[0])
         session.flush()
 
-        assert 0 == self.job_runner.adopt_or_reset_orphaned_tasks(session=session)
+        assert self.job_runner.adopt_or_reset_orphaned_tasks(session=session) == 0
         session.rollback()
 
     def test_adopt_or_reset_orphaned_tasks_stale_scheduler_jobs(self, dag_maker):
@@ -3747,7 +3753,7 @@ class TestSchedulerJob:
 
         num_reset_tis = self.job_runner.adopt_or_reset_orphaned_tasks(session=session)
 
-        assert 1 == num_reset_tis
+        assert num_reset_tis == 1
 
         session.refresh(ti1)
         assert ti1.state is None
@@ -5995,8 +6001,8 @@ class TestSchedulerJob:
             tis = session.query(TaskInstance).all()
 
         dags = self.job_runner.dagbag.dags.values()
-        assert ["test_only_empty_tasks"] == [dag.dag_id for dag in dags]
-        assert 6 == len(tis)
+        assert [dag.dag_id for dag in dags] == ["test_only_empty_tasks"]
+        assert len(tis) == 6
         assert {
             ("test_task_a", "success"),
             ("test_task_b", None),
@@ -6011,7 +6017,7 @@ class TestSchedulerJob:
             if state == "success":
                 assert start_date is not None
                 assert end_date is not None
-                assert 0.0 == duration
+                assert duration == 0.0
             else:
                 assert start_date is None
                 assert end_date is None
@@ -6021,7 +6027,7 @@ class TestSchedulerJob:
         with create_session() as session:
             tis = session.query(TaskInstance).all()
 
-        assert 6 == len(tis)
+        assert len(tis) == 6
         assert {
             ("test_task_a", "success"),
             ("test_task_b", "success"),
@@ -6036,7 +6042,7 @@ class TestSchedulerJob:
             if state == "success":
                 assert start_date is not None
                 assert end_date is not None
-                assert 0.0 == duration
+                assert duration == 0.0
             else:
                 assert start_date is None
                 assert end_date is None