+++ providers/src/airflow/providers/edge/CHANGELOG.rst
@@ -26,6 +26,21 @@
 
 Changelog
 ---------
+0.7.1pre0
+.........
+
+Misc
+~~~~
+
+* ``Edge worker state is sent as 0 to DB if offline or unknown.``
+
+0.7.0pre0
+.........
+
+Misc
+~~~~
+
+* ``Edge worker supports concurrency slots feature so that jobs which need more concurrency blocking other jobs beeing executed on the same worker in parallel.``
 
 0.6.2pre0
 .........